Moved single components to subfolder, img alts, icon titles, share component for single items
