+++ src/app/app-routing.module.ts
@@ -9,9 +9,9 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { HomePageComponent } from './component/home-page/home-page.component';
 import { ResultsComponent } from './component/results/results.component';
-import { SinglePublicationComponent } from './component/single-publication/single-publication.component';
-import { SingleFundingComponent } from './component/single-funding/single-funding.component';
-import { SingleOrganizationComponent } from './component/single-organization/single-organization.component';
+import { SinglePublicationComponent } from './component/single/single-publication/single-publication.component';
+import { SingleFundingComponent } from './component/single/single-funding/single-funding.component';
+import { SingleOrganizationComponent } from './component/single/single-organization/single-organization.component';
 import { VisualisationComponent } from './component/visualisation/visualisation.component';
 import { NewsComponent } from './component/news/news.component';
