+++ src/app/component/results/organizations/organizations.component.html
@@ -26,7 +26,7 @@
   <div *ngFor="let organizations of resultData;">
       <div class="row no-margin single" style="margin: 20px auto!important" *ngFor="let organization of organizations.hits.hits; let i = index">
           <div class="col col-1 icon" style="margin-right: 10px; margin-top: -18px;">
-              <a class="icon" routerLink="/results/organization/{{organization._source.organizationId}}" tabindex="-1">
+              <a class="icon" routerLink="/results/organization/{{organization._source.organizationId}}" tabindex="-1" title="Organisaatioiden tiedon ikoni">
                 <fa-icon [icon]="faIcon"></fa-icon>
               </a>
           </div>
