+++ src/app/component/single/share/share.component.html
@@ -1,0 +1,11 @@
+<aside>
+    <mat-card>
+        <mat-card-title><h2>Jaa</h2></mat-card-title>
+        <div class="social">
+            <a href="#" class="icon" title="Facebook -palvelun logo"><fa-icon [icon]="faFacebook"></fa-icon></a>
+            <a href="#" class="icon" title="Twitter -palvelun logo"><fa-icon [icon]="faTwitter"></fa-icon></a>
+            <a href="#" class="icon" title="LinkedIn -palvelun logo"><fa-icon [icon]="faLinkedin"></fa-icon></a>
+            <a href="#" class="icon" title="Mendeley -palvelun logo"><fa-icon [icon]="faMendeley"></fa-icon></a>
+        </div>
+    </mat-card>
+</aside>
