+++ src/app/component/single/single-funding/single-funding.component.scss
