+++ src/app/component/single/single-organization/single-organization.component.html
@@ -98,17 +98,8 @@
                     </div>
                 </mat-card>
             </aside>
-            <aside>
-                <mat-card>
-                    <mat-card-title><h2>Jaa</h2></mat-card-title>
-                    <div class="social">
-                        <a href="#" class="icon"><fa-icon [icon]="faFacebook"></fa-icon></a>
-                        <a href="#" class="icon"><fa-icon [icon]="faTwitter"></fa-icon></a>
-                        <a href="#" class="icon"><fa-icon [icon]="faLinkedin"></fa-icon></a>
-                        <a href="#" class="icon"><fa-icon [icon]="faMendeley"></fa-icon></a>
-                    </div>
-                </mat-card>
-            </aside>
+            <!-- Social icons -->
+            <app-share></app-share>
             <aside>
                 <mat-card>
                     <mat-card-title><h2>Organisaatiotiedon lähde</h2></mat-card-title>
