+++ src/app/component/single/single-publication/single-publication.component.scss
