+++ src/app/ui/footer/footer.component.html
@@ -9,7 +9,9 @@
   <nav class="footer wrapper">
     <div class="row">
       <div class="col col-3">
-        <a href=""><img class="logo" src="../../../assets/okm_logo.svg"></a>
+        <a href="javascript:void(0)">
+          <img class="logo" src="../../../assets/okm_logo.svg" alt="Opetus- ja kulttuuriministeriön logo, leijonavaakuna vihreällä pohjalla">
+        </a>
       </div>
       <div class="col col-3">
         <ul>
@@ -36,9 +38,9 @@
           <ul>
             <li i18n="@@followUs">Seuraa meitä!</li>
           </ul>
-          <a href="#"><i class="fa fa-facebook-official"></i></a>
-          <a href="#"><i class="fa fa-twitter"></i></a>
-          <a href="#"><i class="fa fa-linkedin-square"></i></a>
+          <a href="#"><i class="fa fa-facebook-official" title="Facebook -palvelun logo"></i></a>
+          <a href="#"><i class="fa fa-twitter" title="Twitter -palvelun logo"></i></a>
+          <a href="#"><i class="fa fa-linkedin-square" title="LinkedIn -palvelun logo"></i></a>
         </div>
       </div>
         <div class="build-info pull-right">
