+++ src/app/ui/footer/footer.component.scss
@@ -37,7 +37,7 @@ img {
 img.logo {
     max-width: 150px;
     height: auto;
-    margin-top: 30px;
+    margin-top: 20px;
 }
 
 .social ul {
