+++ src/app/ui/header/header.component.html
@@ -17,7 +17,7 @@
     <div offClick class="navbar navbar-expand-lg navbar-light row">
         <div class="logo col-9 col-lg-3">
             <a class="navbar-brand" routerLink="">
-                <img src="../../../assets/logo.png" alt="Research.fi">
+                <img src="../../../assets/logo.png" alt="Tutkimustietovarannon portaalin logo, abstrakti ikkuna">
             </a>
             <div class="logo-text d-none d-sm-block"><span i18n="Site slogan @@appSlogan">Ikkuna tutkimukseen Suomessa</span></div>
         </div>
