+++ src/styles.scss
@@ -31,8 +31,8 @@
 @import './app/component/breadcrumb/breadcrumb.component.scss-theme.scss';
 @import './app/component/results/filters/filters-shared.scss';
 @import './app/component/results/filters/filters.theme-shared.scss';
-@import './app/component/single-publication/single-publication.component.scss-theme.scss';
-@import './app/component/single-funding/single-funding.component.scss-theme.scss';
+@import './app/component/single/single-publication/single-publication.component.scss-theme.scss';
+@import './app/component/single/single-funding/single-funding.component.scss-theme.scss';
 
 // Global element color definitions
 @mixin researchfi-components-theme($theme) {
